refactor to match the published api structure, sanity as the api expands
