+++ lib/gowalla.rb
@@ -12,22 +12,6 @@ module Gowalla
 
   VERSION = "0.3.0".freeze
 
-  # config/initializers/gowalla.rb (for instance)
-  #
-  # Gowalla.configure do |config|
-  #   config.api_key = 'api_key'
-  #   config.username = 'username'
-  #   config.password = 'password'
-  # end
-  #
-  # elsewhere
-  #
-  # client = Gowalla::Client.new
-  def self.configure
-    yield self
-    true
-  end
-
   class << self
     attr_accessor :api_key
     attr_accessor :username
@@ -35,11 +19,34 @@ module Gowalla
     attr_accessor :api_secret
     attr_accessor :test_mode
 
+    # config/initializers/gowalla.rb (for instance)
+    #
+    # Gowalla.configure do |config|
+    #   config.api_key = 'api_key'
+    #   config.username = 'username'
+    #   config.password = 'password'
+    # end
+    #
+    # elsewhere
+    #
+    # client = Gowalla::Client.new
+    def configure
+      yield self
+      true
+    end
+
     def test_mode?
       !!self.test_mode
     end
   end
 
+  autoload :Client, 'gowalla/client'
+  autoload :Spots, 'gowalla/spots'
+  autoload :Items, 'gowalla/items'
+  autoload :Users, 'gowalla/users'
+  autoload :Trips, 'gowalla/trips'
+  autoload :Checkins, 'gowalla/checkins'
+
 end
 
-require File.join(directory, 'gowalla', 'client')
+# require File.join(directory, 'gowalla', 'client')
